A hidden input to hold uploaded file reference
After a file is uploaded, we get some sort of data against the uploaded file from the server. It can be a path, URL, or attachment ID. In this commit, that returned ref is stored in a hidden input for each uploaded file preview. This will help perform ops w.r.t the uploaded file.